Fix styling errors and add interesting css
This commit doesn't change any of the basic content of the website,
however, it adds a lot of functionality that I previously couldn't do.
There is a typewriter effect that mostly works and the image sliding
that I recently made.

I also worked to fix the sizing of the website so that it will fit
with any computer screen rather than just a macbook. I did this
primarily us vh&vw and further employing Flexbox.